remove all `controller` usage from declaration files etc.
- also refactor the structure of declaration files for new platforms (former controllers)
