+++ @types/platforms/minecraft-platform.d.ts
@@ -1,4 +1,4 @@
-import { Platform } from "../platform";
+import { Platform } from "./template";
 
 // currently unused
 export declare class MinecraftPlatform extends Platform {
